Header based on user type (#38)
* Header based on user type

* Changed business role

* Chenged customer and superuser roles

* подредачи хедер

---------
