+++ src/app/dashboard/home/page.tsx
@@ -1,0 +1,9 @@
+'use client'
+
+export default function Home() {
+  return (
+    <>
+      Home
+    </>
+  );
+}
