--- src/app/home/page.tsx
@@ -1,12 +1,0 @@
-'use client'
-
-import BusinessHeader from '@/components/heading/BusinessHeader'
-
-export default function Home() {
-  return (
-    <div className='container'>
-      <BusinessHeader />
-      Home
-    </div>
-  );
-}
